updated A to Q1
Corrected technical ambiguity, simply false misinformation, spelling mistakes and grammar (hope i didn't leave any:) )
This question is super important and is the first one for a very good reason.

the first two paragraphs were (IMO) completely un-called for, if a question starts with "What was the impetus?" the answer starts with "The impetus was" and not with changing the merit of the question. It is the responsibility of the CFTC to ask questions and it's our willingness to answer them that will help, changing the question and adding un-called for lines does not help.

By technical ambiguity and simply false misinformation i refer to Bitcoin lacking programmability and that being the key difference.
This is factually incorrect, fixed the answer by noting Turing completeness as the key change and providing links for what Turing completeness and conditional branching are.