Update what-is-ethereum.md
I thought this could be tuned up a little bit. Sounded a little bit too hype-ish. Adjusted to replace "anyone" to "developers" to provide a more realistic expectation of coding ability. Also, removed the "exactly as coded" due to the PTSD it's giving me from the DAO and Parity Multi-sigs. Technically true, I think that isn't as much of a selling point as it use to be versus execution in a trustless environment with no gate keeper.